Auth corrections

+++ app/Http/Controllers/BooksController.php
@@ -5,13 +5,12 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Book;
+use Illuminate\Support\Facades\Auth;
+use phpDocumentor\Reflection\Types\Self_;
 use Session;
 
 class BooksController extends Controller
 {
-    public function __construct() {
-      $this->middleware('auth');
-    }
 
     /**
      * Display a listing of the books.
@@ -20,11 +19,16 @@ class BooksController extends Controller
      */
     public function index() {
 
-      // Variable to store all the books from the database.
-      $books = Book::orderBy('id', 'desc')->paginate(6);
+      if (Auth::check()) {
+        // Variable to store all the books from the database.
+        $books = Book::orderBy('id', 'desc')->paginate(6);
 
-      // Return the view.
-      return view('books.index')->withBooks($books);
+        // Return the view.
+        return view('books.index')->withBooks($books);
+      }
+      else {
+        return (view('auth.login'));
+      }
     }
 
     /**
@@ -34,8 +38,13 @@ class BooksController extends Controller
      */
     public function create() {
 
-      //Returns the view that creates a new book.
-      return view('books.create');
+      if (Auth::check()) {
+        //Returns the view that creates a new book.
+        return view('books.create');
+      }
+      else {
+        return (view('auth.login'));
+      }
     }
 
     /**
@@ -46,28 +55,30 @@ class BooksController extends Controller
      */
     public function store(Request $request) {
 
-      // Validate the data.
-      $this->validate($request, array(
-        'isbn' => 'required|max:13',
-        'title' => 'required|max:255',
-        'description' => 'required',
-        'author' => 'required',
-        'category' => 'required',
-      ));
+      if (Auth::check()) {
+        // Validate the data.
+        $this->validate($request, array(
+          'isbn' => 'required|max:13',
+          'title' => 'required|max:255',
+          'description' => 'required',
+          'author' => 'required',
+          'category' => 'required',
+        ));
 
-      // Store the book in the database.
-      $book = new Book;
-      $book->isbn = $request->isbn;
-      $book->title = $request->title;
-      $book->description = $request->description;
-      $book->author = $request->author;
-      $book->category = $request->category;
-      $book->save();
+        // Store the book in the database.
+        $book = new Book;
+        $book->isbn = $request->isbn;
+        $book->title = $request->title;
+        $book->description = $request->description;
+        $book->author = $request->author;
+        $book->category = $request->category;
+        $book->save();
 
-      Session::flash('success', 'Your book saved successfully!');
+        Session::flash('success', 'Your book saved successfully!');
 
-      // Redirect to another page.
-      return redirect()->route('books.show', $book->id);
+        // Redirect to another page.
+        return redirect()->route('books.show', $book->id);
+      }
 
     }
 
@@ -92,11 +103,16 @@ class BooksController extends Controller
      */
     public function edit($id) {
 
-      // Find the post in the database.
-      $book = Book::find($id);
+      if (Auth::check()) {
+        // Find the post in the database.
+        $book = Book::find($id);
 
-      // Return the view.
-      return view('books.edit')->withBook($book);
+        // Return the view.
+        return view('books.edit')->withBook($book);
+      }
+      else {
+        return (view('auth.login'));
+      }
 
     }
 
@@ -109,29 +125,31 @@ class BooksController extends Controller
      */
     public function update(Request $request, $id) {
 
-      // Validate the data.
-      $book = Book::find($id);
-      $this->validate($request, array(
-        'isbn' => 'required|max:13',
-        'title' => 'required|max:255',
-        'description' => 'required',
-        'author' => 'required',
-        'category' => 'required',
-      ));
+      if (Auth::check()) {
+        // Validate the data.
+        $book = Book::find($id);
+        $this->validate($request, array(
+          'isbn' => 'required|max:13',
+          'title' => 'required|max:255',
+          'description' => 'required',
+          'author' => 'required',
+          'category' => 'required',
+        ));
 
-      // Save the book to database.
-      $book->isbn = $request->input('isbn');
-      $book->title = $request->input('title');
-      $book->description = $request->input('description');
-      $book->author = $request->input('author');
-      $book->category = $request->input('category');
-      $book->save();
+        // Save the book to database.
+        $book->isbn = $request->input('isbn');
+        $book->title = $request->input('title');
+        $book->description = $request->input('description');
+        $book->author = $request->input('author');
+        $book->category = $request->input('category');
+        $book->save();
 
-      // Show a success message.
-      Session::flash('success', 'This book successfully edited.');
+        // Show a success message.
+        Session::flash('success', 'This book successfully edited.');
 
-      // Redirect to book page.
-      return redirect()->route('books.show', $book->id);
+        // Redirect to book page.
+        return redirect()->route('books.show', $book->id);
+      }
 
     }
 
@@ -143,13 +161,18 @@ class BooksController extends Controller
      */
     public function destroy($id) {
 
-      // Find the book to delete.
-      $book = Book::find($id);
+      if (Auth::check()) {
+        // Find the book to delete.
+        $book = Book::find($id);
 
-      // Delete the book.
-      $book->delete();
+        // Delete the book.
+        $book->delete();
 
-      Session::flash('success', 'The book was deleted successfully');
-      return redirect()->route('books.index');
+        Session::flash('success', 'The book was deleted successfully');
+        return redirect()->route('books.index');
+      }
+      else {
+        return (view('auth.login'));
+      }
     }
 }
